Adds url-as-a-graph feature to dashboards
Before this commit, when writing a dashboard and including a graph in
it, we had to:

1. create a datasource metadata file on top of a datafile
2. create a graph metadata file accessing that datasource

But most of the time we just wanted to graph all available columns from
a datafile.  This change allows one to simply specify a datafile url
instead of a graph id.  Limn fetches the datafile, infers a datasource
from it, and builds a graph of all the metrics in the datasource.
Caveat: DO NOT add multiple datafiles with the same filename to the same
dashboard, unless they're exactly the same (because Limn will cache the
datasource and serve it both times).